feat: do not conflict if server and client previously had same value
Previously, dropped network packets will result in a sync conflict the next time the note successfully reaches the server round trip.

Solution: When the server reports it is unable to save an item due to potential conflict and returns it to the client, the client will:

1. Check the value of the local item before the local change via a new

2. If previousRevision.content === serverValue.content, this means the change that was rejected by the server is in fact a legitimate change, because the value the client had previously matched with the server's, and this new change is being built on top of that state, and should therefore be chosen as the winner, with no need for a conflict.
