+++ packages/snjs/lib/services/sync/account/response_resolver.ts
@@ -5,6 +5,7 @@ import { PayloadSource } from '@Payloads/sources';
 import { ImmutablePayloadCollection } from '@Protocol/collection/payload_collection';
 import { ImmutablePayloadCollectionSet } from '@Protocol/collection/collection_set';
 import { CopyPayload } from '@Payloads/generator';
+import { HistoryMap } from '@Lib/services/history/history_map';
 
 /**
  * Given a remote sync response, the resolver applies the incoming changes on top
@@ -13,18 +14,15 @@ import { CopyPayload } from '@Payloads/generator';
  * offers the 'recommended' new global state given a sync response and a current base state.
  */
 export class SyncResponseResolver {
-  private response: SyncResponse;
-  private baseCollection: ImmutablePayloadCollection;
   private relatedCollectionSet: ImmutablePayloadCollectionSet;
 
   constructor(
-    response: SyncResponse,
+    private response: SyncResponse,
     decryptedResponsePayloads: PurePayload[],
-    baseCollection: ImmutablePayloadCollection,
-    payloadsSavedOrSaving: PurePayload[]
+    private baseCollection: ImmutablePayloadCollection,
+    payloadsSavedOrSaving: PurePayload[],
+    private historyMap: HistoryMap
   ) {
-    this.response = response;
-    this.baseCollection = baseCollection;
     this.relatedCollectionSet = new ImmutablePayloadCollectionSet([
       ImmutablePayloadCollection.WithPayloads(
         decryptedResponsePayloads,
@@ -37,7 +35,9 @@ export class SyncResponseResolver {
     ]);
   }
 
-  public async collectionsByProcessingResponse() {
+  public async collectionsByProcessingResponse(): Promise<
+    ImmutablePayloadCollection[]
+  > {
     const collections = [];
 
     const collectionRetrieved = await this.collectionByProcessingPayloads(
@@ -82,7 +82,7 @@ export class SyncResponseResolver {
   private async collectionByProcessingPayloads(
     payloads: PurePayload[],
     source: PayloadSource
-  ) {
+  ): Promise<ImmutablePayloadCollection> {
     const collection = ImmutablePayloadCollection.WithPayloads(
       payloads,
       source
@@ -92,7 +92,8 @@ export class SyncResponseResolver {
     const delta = new deltaClass(
       this.baseCollection,
       collection,
-      this.relatedCollectionSet
+      this.relatedCollectionSet,
+      this.historyMap,
     );
     const resultCollection = await delta.resultingCollection();
     const updatedDirtyPayloads = resultCollection.all().map((payload) => {
@@ -108,7 +109,7 @@ export class SyncResponseResolver {
     );
   }
 
-  private finalDirtyStateForPayload(payload: PurePayload) {
+  private finalDirtyStateForPayload(payload: PurePayload): boolean | undefined {
     const current = this.baseCollection.find(payload.uuid!);
     /**
      * `current` can be null in the case of new
