+++ test/test.html
@@ -67,7 +67,7 @@
   <script type="module" src="sync_tests/offline.test.js"></script>
   <script type="module" src="sync_tests/notes_tags.test.js"></script>
   <script type="module" src="sync_tests/online.test.js"></script>
-  <script type="module" src="sync_tests/duplication.test.js"></script>
+  <script type="module" src="sync_tests/conflicting.test.js"></script>
   <script type="module" src="sync_tests/discordance.test.js"></script>
   <script type="module" src="auth.test.js"></script>
   <script type="module" src="device_auth.test.js"></script>
